feat: add frontend handling
